cleaned up pyinstaller stuff
