+++ pyausaxs/__init__.py
@@ -1,5 +1,3 @@
-"""pyausaxs package init"""
-
 from .wrapper import AUSAXS, AUSAXSManualFit, ausaxs, create_ausaxs
 
 __all__ = ["AUSAXS", "AUSAXSManualFit", "ausaxs", "create_ausaxs"]
