+++ pyausaxs/__pyinstaller/pyinstaller.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+def get_hook_dirs():
+    return [str(Path(__file__).parent)]
